+++ src/constants/routes-constant.js
@@ -13,10 +13,11 @@ const NAME = Object.freeze({
   BLOCKS_LIST: 'blocksList',
   BLOCKS_DETAIL: 'blocksDetail',
   DEX: 'dex',
-  DEX_LIST: 'dexList',
-  DEX_DETAIL: 'dexDetail',
-  DEX_DETAIL_DASHBOARD: 'dexDetailDashboard',
-  DEX_DETAIL_SWAP: 'dexDetailSwap',
+  POOLS: 'pools',
+  POOLS_LIST: 'poolsList',
+  POOLS_DETAIL: 'poolsDetail',
+  POOLS_DETAIL_DASHBOARD: 'poolsDetailDashboard',
+  POOLS_DETAIL_SWAP: 'poolsDetailSwap',
   HOME: 'home',
   EXCHANGE_RATE: 'exchangeRate',
   PROPOSALS: 'proposals',
@@ -51,10 +52,11 @@ const PATH = Object.freeze({
   BLOCKS_LIST: 'list',
   BLOCKS_DETAIL: 'detail/:id',
   DEX: '/dex',
-  DEX_LIST: 'list',
-  DEX_DETAIL: 'detail/:id',
-  DEX_DETAIL_DASHBOARD: 'dashboard',
-  DEX_DETAIL_SWAP: 'swap',
+  POOLS: 'pools',
+  POOLS_LIST: 'list',
+  POOLS_DETAIL: 'detail/:id',
+  POOLS_DETAIL_DASHBOARD: 'dashboard',
+  POOLS_DETAIL_SWAP: 'swap',
   HOME: '/home',
   EXCHANGE_RATE: '/exchange-rate',
   PROPOSALS: '/proposals',
@@ -88,10 +90,11 @@ const TITLE = Object.freeze({
   BLOCKS_LIST: 'Blocks List',
   BLOCKS_DETAIL: 'Blocks Detail',
   DEX: 'dex',
-  DEX_LIST: 'DEX List',
-  DEX_DETAIL: 'DEX Detail',
-  DEX_DETAIL_DASHBOARD: 'DEX Detail Dashboard',
-  DEX_DETAIL_SWAP: 'DEX Detail Swap',
+  POOLS: 'pools',
+  POOLS_LIST: 'Pools List',
+  POOLS_DETAIL: 'Pools Detail',
+  POOLS_DETAIL_DASHBOARD: 'Pools Detail Dashboard',
+  POOLS_DETAIL_SWAP: 'Pools Detail Swap',
   HOME: 'Home',
   EXCHANGE_RATE: 'Exchange Rate',
   PROPOSALS: 'Proposals',
