+++ src/router/index.js
@@ -7,10 +7,10 @@ import store from '@/store';
 import accountRoute from './routes/account-route';
 import assetsRoute from './routes/assets-route';
 import blocksRoute from './routes/blocks-route';
-import dexRoute from './routes/dex-route';
 import exchangeRateRoute from './routes/exchangeRate-route';
 import homeRoute from './routes/home-route';
 import issueRoutes from './routes/issue-routes';
+import poolsRoute from './routes/pools-route';
 import proposalsRoute from './routes/proposals-route';
 import statsRoute from './routes/stats-route';
 import transactionsRoute from './routes/transactions-route';
@@ -31,9 +31,9 @@ routes.push(
   accountRoute,
   assetsRoute,
   blocksRoute,
-  dexRoute,
   exchangeRateRoute,
   homeRoute,
+  poolsRoute,
   proposalsRoute,
   statsRoute,
   transactionsRoute,
