+++ src/store/pools/list/__tests__/mutations.test.js
@@ -1,0 +1,44 @@
+import { initState } from '../index';
+import mutations from '../mutations';
+
+describe('store/poolsList/mutations', () => {
+  let state = {};
+
+  beforeEach(() => {
+    state = initState();
+  });
+
+  test('mutations.reset', () => {
+    state.error = 'error';
+    state.isLoading = true;
+    state.contracts = [{ id: 1 }];
+
+    const expected = initState();
+
+    mutations.reset(state);
+
+    expect(state).toStrictEqual(expected);
+  });
+
+  test('mutations.setError', () => {
+    const payload = 'error';
+
+    mutations.setError(state, payload);
+
+    expect(state.error).toBe(payload);
+  });
+
+  test('mutations.setLoading', () => {
+    mutations.setLoading(state, true);
+
+    expect(state.isLoading).toBe(true);
+  });
+
+  test('mutations.addContract', () => {
+    const payload = { id: 1 };
+
+    mutations.addContract(state, payload);
+
+    expect(state.contracts).toStrictEqual([payload]);
+  });
+});
