--- src/views/dex/DexRouterDetail.vue
@@ -1,13 +1,0 @@
-<template>
-  <div>
-    <transition name="route">
-      <router-view />
-    </transition>
-  </div>
-</template>
-
-<script>
-export default {
-  name: 'DexRouterDetail',
-};
-</script>
