+++ src/views/pools/PoolsRouterDetail.vue
@@ -1,0 +1,13 @@
+<template>
+  <div>
+    <transition name="route">
+      <router-view />
+    </transition>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'PoolsRouterDetail',
+};
+</script>
