+++ src/views/pools/__tests__/PoolsViewList.test.js
@@ -1,0 +1,69 @@
+import { createLocalVue, shallowMount } from '@vue/test-utils';
+import Vuex from 'vuex';
+import PoolsViewList from '../PoolsViewList.vue';
+
+const localVue = createLocalVue();
+
+describe('views/pools/PoolsViewList.vue', () => {
+  const actions = {
+    initPoolsList: jest.fn(),
+  };
+  const mockStore = new Vuex.Store({
+    modules: {
+      poolsList: {
+        namespaced: true,
+        actions,
+      },
+    },
+  });
+  const mocks = {
+    $store: mockStore,
+    $t: (msg) => msg,
+  };
+
+  test('if loading indicator is displayed', () => {
+    const wrapper = shallowMount(PoolsViewList, {
+      localVue,
+      mocks,
+      computed: {
+        error: () => null,
+        isLoading: () => true,
+      },
+    });
+
+    expect(wrapper.find('[data-test="loading"]').exists()).toBe(true);
+    expect(wrapper.find('[data-test="error"]').exists()).toBe(false);
+    expect(wrapper.find('[data-test="content"]').exists()).toBe(false);
+  });
+
+  test('if message error is displayed', () => {
+    const error = Error('message');
+    const wrapper = shallowMount(PoolsViewList, {
+      localVue,
+      mocks,
+      computed: {
+        error: () => error,
+        isLoading: () => false,
+      },
+    });
+
+    expect(wrapper.find('[data-test="loading"]').exists()).toBe(false);
+    expect(wrapper.find('[data-test="error"]').exists()).toBe(true);
+    expect(wrapper.find('[data-test="content"]').exists()).toBe(false);
+  });
+
+  test('if content is displayed', () => {
+    const wrapper = shallowMount(PoolsViewList, {
+      localVue,
+      mocks,
+      computed: {
+        error: () => null,
+        isLoading: () => false,
+      },
+    });
+
+    expect(wrapper.find('[data-test="loading"]').exists()).toBe(false);
+    expect(wrapper.find('[data-test="error"]').exists()).toBe(false);
+    expect(wrapper.find('[data-test="content"]').exists()).toBe(true);
+  });
+});
